HomeWork lab1 Game Update3 (Warriors Ready)

+++ 321labs/LabGame/Base/IScan.cs
@@ -6,7 +6,7 @@ namespace _321labs.LabGame.Base
 {
     public interface IScan
     {
-        public abstract int Sense { get; set; }
+        public abstract float Sense { get; set; }
         public abstract List<Unit> ScanAround();
     }
 }
